Moved documentation utils to Aws::Api::Docs module.
