+++ doc-src/plugins/apis.rb
@@ -11,6 +11,6 @@ YARD::Templates::Engine.register_template_path(File.join(File.dirname(__FILE__),
 YARD::Parser::SourceParser.after_parse_list do
   Aws.load_all_services
   Aws.service_added do |_, svc_module, options|
-    Aws::Api::Documenter.new(svc_module).apply
+    Aws::Api::Docs::Builder.document(svc_module)
   end
 end
